Stub share and notification services on web
