+++ lib/services/notification_service_web.dart
@@ -1,0 +1,8 @@
+// lib/services/notification_service_web.dart
+import 'package:flutter/material.dart';
+
+class NotificationService {
+  static Future<void> init() async {}
+  void initTimeZones() {}
+  Future<void> scheduleDailyNotifications(List<TimeOfDay> times) async {}
+}
